feat: `actors pull`
